requirements.txt
mss
opencv-python
pillow
numpy
pywin32